requirements.txt
dhooks
plexapi
pyyaml
schedule
requests